Update services/requirements to merge in versioning
This removes the Versioned module for these requirement classes and
consolidates the files.
