+++ app/services/requirements/edit_page_checker.rb
@@ -2,16 +2,17 @@
 
 module Requirements
   class EditPageChecker
-    attr_reader :document
+    attr_reader :edition, :revision
 
-    def initialize(document)
-      @document = document
+    def initialize(edition, revision = nil)
+      @edition = edition
+      @revision = revision || edition.revision
     end
 
     def pre_preview_issues
       issues = []
-      issues += PathChecker.new(document).pre_preview_issues.to_a
-      issues += ContentChecker.new(document).pre_preview_issues.to_a
+      issues += PathChecker.new(edition, revision).pre_preview_issues.to_a
+      issues += ContentChecker.new(edition, revision).pre_preview_issues.to_a
       CheckerIssues.new(issues)
     end
   end
