+++ app/services/requirements/path_checker.rb
@@ -2,17 +2,18 @@
 
 module Requirements
   class PathChecker
-    attr_reader :document
+    attr_reader :edition, :revision
 
-    def initialize(document)
-      @document = document
+    def initialize(edition, revision = nil)
+      @edition = edition
+      @revision = revision || edition.revision
     end
 
     def pre_preview_issues
       issues = []
 
       begin
-        if document.document_type.check_path_conflict && base_path_conflict?
+        if edition.document_type.check_path_conflict && base_path_conflict?
           issues << Issue.new(:title, :conflict)
         end
       rescue GdsApi::BaseError => e
@@ -26,13 +27,13 @@ module Requirements
 
     def base_path_conflict?
       base_path_owner = GdsApi.publishing_api_v2.lookup_content_id(
-        base_path: document.base_path,
+        base_path: revision.base_path,
         with_drafts: true,
         exclude_document_types: [],
         exclude_unpublishing_types: [],
       )
 
-      base_path_owner && base_path_owner != document.content_id
+      base_path_owner && base_path_owner != edition.content_id
     end
   end
 end
