+++ spec/services/requirements/image_revision_checker_spec.rb
@@ -1,16 +1,16 @@
 # frozen_string_literal: true
 
-RSpec.describe Versioned::Requirements::ImageRevisionChecker do
+RSpec.describe Requirements::ImageRevisionChecker do
   describe "#pre_preview_issues" do
     it "returns no issues if there are none" do
-      image_revision = build :versioned_image_revision, alt_text: "something"
-      issues = Versioned::Requirements::ImageRevisionChecker.new(image_revision).pre_preview_issues
+      image_revision = build :image_revision, alt_text: "something"
+      issues = Requirements::ImageRevisionChecker.new(image_revision).pre_preview_issues
       expect(issues.items).to be_empty
     end
 
     it "returns an issue if there is no alt text" do
-      image_revision = build :versioned_image_revision
-      issues = Versioned::Requirements::ImageRevisionChecker.new(image_revision).pre_preview_issues
+      image_revision = build :image_revision
+      issues = Requirements::ImageRevisionChecker.new(image_revision).pre_preview_issues
 
       form_message = issues.items_for(:alt_text).first[:text]
       expect(form_message).to eq(I18n.t!("requirements.alt_text.blank.form_message"))
@@ -21,9 +21,9 @@ RSpec.describe Versioned::Requirements::ImageRevisionChecker do
     end
 
     it "returns an issue if the alt text is too long" do
-      max_length = Versioned::Requirements::ImageRevisionChecker::ALT_TEXT_MAX_LENGTH
-      image_revision = build :versioned_image_revision, alt_text: "a" * (max_length + 1)
-      issues = Versioned::Requirements::ImageRevisionChecker.new(image_revision).pre_preview_issues
+      max_length = Requirements::ImageRevisionChecker::ALT_TEXT_MAX_LENGTH
+      image_revision = build :image_revision, alt_text: "a" * (max_length + 1)
+      issues = Requirements::ImageRevisionChecker.new(image_revision).pre_preview_issues
 
       form_message = issues.items_for(:alt_text).first[:text]
       expect(form_message)
@@ -39,9 +39,9 @@ RSpec.describe Versioned::Requirements::ImageRevisionChecker do
     end
 
     it "returns an issue if the caption is too long" do
-      max_length = Versioned::Requirements::ImageRevisionChecker::CAPTION_MAX_LENGTH
-      image_revision = build :versioned_image_revision, caption: "a" * (max_length + 1)
-      issues = Versioned::Requirements::ImageRevisionChecker.new(image_revision).pre_preview_issues
+      max_length = Requirements::ImageRevisionChecker::CAPTION_MAX_LENGTH
+      image_revision = build :image_revision, caption: "a" * (max_length + 1)
+      issues = Requirements::ImageRevisionChecker.new(image_revision).pre_preview_issues
 
       form_message = issues.items_for(:caption).first[:text]
       expect(form_message)
@@ -57,9 +57,9 @@ RSpec.describe Versioned::Requirements::ImageRevisionChecker do
     end
 
     it "returns an issue if the credit is too long" do
-      max_length = Versioned::Requirements::ImageRevisionChecker::CREDIT_MAX_LENGTH
-      image_revision = build :versioned_image_revision, credit: "a" * (max_length + 1)
-      issues = Versioned::Requirements::ImageRevisionChecker.new(image_revision).pre_preview_issues
+      max_length = Requirements::ImageRevisionChecker::CREDIT_MAX_LENGTH
+      image_revision = build :image_revision, credit: "a" * (max_length + 1)
+      issues = Requirements::ImageRevisionChecker.new(image_revision).pre_preview_issues
 
       form_message = issues.items_for(:credit).first[:text]
       expect(form_message)
