+++ spec/services/requirements/path_checker_spec.rb
@@ -5,8 +5,8 @@ RSpec.describe Requirements::PathChecker do
     context "when the format does not check paths" do
       it "returns no issues" do
         document_type = build :document_type
-        document = build :document, document_type_id: document_type.id
-        issues = Requirements::PathChecker.new(document).pre_preview_issues
+        edition = build :edition, document_type_id: document_type.id
+        issues = Requirements::PathChecker.new(edition).pre_preview_issues
         expect(issues.items).to be_empty
       end
     end
@@ -14,31 +14,45 @@ RSpec.describe Requirements::PathChecker do
     context "when the Publishing API is available" do
       it "returns no issues for unreserved paths" do
         document_type = build :document_type, check_path_conflict: true
-        document = build :document, document_type_id: document_type.id
-        publishing_api_has_lookups(document.base_path => nil)
-        issues = Requirements::PathChecker.new(document).pre_preview_issues
+        edition = build :edition, document_type_id: document_type.id
+        publishing_api_has_lookups(edition.base_path => nil)
+        issues = Requirements::PathChecker.new(edition).pre_preview_issues
         expect(issues.items).to be_empty
       end
 
       it "returns no issues if the document owns the path" do
         document_type = build :document_type, check_path_conflict: true
-        document = build :document, document_type_id: document_type.id
-        publishing_api_has_lookups(document.base_path => document.content_id)
-        issues = Requirements::PathChecker.new(document).pre_preview_issues
+        edition = build :edition, document_type_id: document_type.id
+        publishing_api_has_lookups(edition.base_path => edition.content_id)
+        issues = Requirements::PathChecker.new(edition).pre_preview_issues
         expect(issues.items).to be_empty
       end
 
       it "returns an issue if the base_path conflicts" do
         document_type = build :document_type, check_path_conflict: true
-        document = build :document, document_type_id: document_type.id
-        publishing_api_has_lookups(document.base_path => SecureRandom.uuid)
-        issues = Requirements::PathChecker.new(document).pre_preview_issues
+        edition = build :edition, document_type_id: document_type.id
+        publishing_api_has_lookups(edition.base_path => SecureRandom.uuid)
+        issues = Requirements::PathChecker.new(edition).pre_preview_issues
 
         form_message = issues.items_for(:title).first[:text]
         expect(form_message).to eq(I18n.t!("requirements.title.conflict.form_message"))
 
         summary_message = issues.items_for(:title, style: "summary").first[:text]
         expect(summary_message).to eq(I18n.t!("requirements.title.conflict.summary_message"))
+      end
+
+      it "can check a revision" do
+        document_type = build :document_type, check_path_conflict: true
+        edition = build :edition, document_type_id: document_type.id
+        revision = build :revision
+        publishing_api_has_lookups(edition.base_path => nil,
+                                   revision.base_path => SecureRandom.uuid)
+        issues = Requirements::PathChecker.new(edition).pre_preview_issues
+        expect(issues.items).to be_empty
+
+        issues = Requirements::PathChecker.new(edition, revision)
+                                                     .pre_preview_issues
+        expect(issues.items).not_to be_empty
       end
     end
 
@@ -49,8 +63,8 @@ RSpec.describe Requirements::PathChecker do
 
       it "returns no issues (ignore exception)" do
         document_type = build :document_type, check_path_conflict: true
-        document = build :document, document_type_id: document_type.id
-        issues = Requirements::PathChecker.new(document).pre_preview_issues
+        edition = build :edition, document_type_id: document_type.id
+        issues = Requirements::PathChecker.new(edition).pre_preview_issues
         expect(issues.items_for(:title)).to be_empty
       end
     end
