+++ spec/services/requirements/topic_checker_spec.rb
@@ -5,8 +5,8 @@ RSpec.describe Requirements::TopicChecker do
 
   describe "#pre_publish_issues" do
     it "returns no issues if there are none" do
-      document = build :document, :publishable
-      issues = Requirements::TopicChecker.new(document).pre_publish_issues
+      edition = build :edition, :publishable
+      issues = Requirements::TopicChecker.new(edition.document).pre_publish_issues
       expect(issues.items).to be_empty
     end
